Moved DI package and minors fixes
