+++ app/src/androidTest/java/com/prestosoftware/test/rappi/runner/AndroidJunitTestRunner.kt
@@ -3,12 +3,12 @@ package com.prestosoftware.test.rappi.runner
 import android.app.Application
 import android.content.Context
 import androidx.test.runner.AndroidJUnitRunner
-import com.prestosoftware.test.rappi.TestRappiApplication
+import com.prestosoftware.test.TestApplication
 
 @Suppress("unused")
 class AndroidJunitTestRunner : AndroidJUnitRunner() {
   @Throws(InstantiationException::class, IllegalAccessException::class, ClassNotFoundException::class)
   override fun newApplication(cl: ClassLoader, className: String, context: Context): Application {
-    return super.newApplication(cl, TestRappiApplication::class.java.name, context)
+    return super.newApplication(cl, TestApplication::class.java.name, context)
   }
 }
