+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".TestRappiApplication"
+        android:name="com.prestosoftware.test.TestApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.prestosoftware.test.reign.ui.post.detail.PostDetailActivity"
+            android:name="com.prestosoftware.test.reign.ui.detail.PostDetailActivity"
             android:exported="false" />
         <activity
             android:name="com.prestosoftware.test.reign.ui.post.PostActivity"
