+++ app/src/main/java/com/prestosoftware/test/di/AppComponent.kt
@@ -1,4 +1,4 @@
-package com.prestosoftware.test.rappi.di
+package com.prestosoftware.test.di
 
 import android.app.Application
 import dagger.BindsInstance
