+++ app/src/main/java/com/prestosoftware/test/di/MainActivityFragmentModule.kt
@@ -1,4 +1,4 @@
-package com.prestosoftware.test.rappi.di
+package com.prestosoftware.test.di
 
 import com.prestosoftware.test.rappi.ui.movie.upcoming.MovieUpcomingListFragment
 import com.prestosoftware.test.rappi.ui.movie.popular.MoviePopularListFragment
