+++ app/src/main/java/com/prestosoftware/test/di/PersistenceModule.kt
@@ -1,9 +1,9 @@
-package com.prestosoftware.test.rappi.di
+package com.prestosoftware.test.di
 
 import android.app.Application
 import androidx.annotation.NonNull
 import androidx.room.Room
-import com.prestosoftware.test.rappi.TestRappiApplication
+import com.prestosoftware.test.TestApplication
 import com.prestosoftware.test.rappi.data.db.MovieDao
 import com.prestosoftware.test.reign.data.dao.PostDao
 import com.prestosoftware.test.rappi.data.db.TestDb
@@ -18,7 +18,7 @@ class PersistenceModule {
   @Singleton
   fun provideDatabase(@NonNull application: Application): TestDb {
     return Room
-      .databaseBuilder(application, TestDb::class.java, TestRappiApplication.BD_NAME)
+      .databaseBuilder(application, TestDb::class.java, TestApplication.BD_NAME)
       .allowMainThreadQueries()
       .build()
   }
