+++ app/src/main/java/com/prestosoftware/test/di/ViewModelKey.kt
@@ -1,4 +1,4 @@
-package com.prestosoftware.test.rappi.di
+package com.prestosoftware.test.di
 
 import androidx.lifecycle.ViewModel
 import dagger.MapKey
