+++ app/src/main/res/layout/activity_post_detail.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.prestosoftware.test.reign.ui.post.detail.PostDetailActivity">
+    tools:context="com.prestosoftware.test.reign.ui.detail.PostDetailActivity">
 
     <WebView
         android:id="@+id/web_view"
